newest (fix some)

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: kialvare <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/28 11:06:00 by kialvare          #+#    #+#             */
-/*   Updated: 2016/09/28 11:09:07 by kialvare         ###   ########.fr       */
+/*   Updated: 2016/10/02 10:10:18 by kialvare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,17 +14,17 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int i;
-	char *d;
-	const char *s;
+	size_t		i;
+	char 		*d;
+	const char 	*s;
 
 	i = 0;
 	d = dest;
 	s = src;
-	while (i < (int)n)
+	while (i < n)
 	{
 		d[i] = s[i];
 		i++;
 	}
-	return(dest);
+	return (dest);
 }
